Current code that was tested on 2017-11-17 -- not reading correctly

+++ src/adc.h
@@ -46,7 +46,7 @@ void write_ADC_register(uint8_t register_addr, uint32_t data);
 
 void select_ADC_channel(uint8_t channel_num);
 uint32_t read_ADC_channel(uint8_t channel_num);
-double convert_ADC_reading(uint8_t ADC_reading, uint8_t pga_gain);
+double convert_ADC_reading(uint32_t ADC_reading, uint8_t pga_gain);
 
 uint8_t num_register_bytes(uint8_t register_addr);
 void set_PGA(uint8_t gain);
